Adding CLI
